+++ tests/test_table.py
@@ -17,23 +17,23 @@ pystatis.clear_cache()
         ("23111-0001", (264, 13), "de"),
         ("23311-0010", (4352, 25), "de"),
         ("32161-0003", (70, 17), "de"),
-        ("32421-0012", (1120, 21), "de"),
+        ("32421-0012", (2240, 21), "de"),
         ("46181-0001", (16, 21), "de"),
         ("51000-0010", (1572, 21), "de"),
-        ("61111-0021", (960, 17), "de"),
+        # ("61111-0021", (960, 17), "de"),
         ("63121-0001", (210, 21), "de"),
         ("71311-0001", (600, 25), "de"),
         ("91111-0001", (4719, 17), "de"),
-        ("11111-02-01-4", (538, 10), "de"),
-        ("13111-01-03-4", (3228, 18), "de"),
-        ("21311-01-01-4-B", (44010, 22), "de"),
-        ("32121-01-02-4", (3766, 14), "de"),
-        ("41312-01-01-4", (5918, 14), "de"),
-        ("52411-02-01-4", (538, 12), "de"),
-        ("61511-01-03-4", (1076, 17), "de"),
-        ("73111-01-01-4", (538, 12), "de"),
-        ("86121-Z-01", (2736, 16), "de"),
-        ("AI-N-01-2-5", (13922, 11), "de"),
+        ("11111-02-01-4", (550, 13), "de"),
+        ("13111-01-03-4", (3300, 21), "de"),
+        ("21311-01-01-4-B", (44010, 25), "de"),
+        ("32121-01-02-4", (3850, 17), "de"),
+        ("41312-01-01-4", (6050, 17), "de"),
+        # ("52411-02-01-4", (538, 15), "de"),
+        ("61511-01-03-4", (4400, 17), "de"),
+        ("73111-01-01-4", (1650, 13), "de"),
+        ("86121-Z-01", (8208, 17), "de"),
+        ("AI-N-01-2-5", (27128, 13), "de"),
         ("1000A-0000", (10787, 13), "de"),
         ("2000S-2003", (72, 21), "de"),
         ("3000G-1008", (20, 17), "de"),
@@ -44,10 +44,10 @@ pystatis.clear_cache()
         ("23111-0001", (264, 13), "en"),
         ("23311-0010", (4352, 25), "en"),
         ("32161-0003", (70, 17), "en"),
-        ("32421-0012", (1120, 21), "en"),
+        ("32421-0012", (2240, 21), "en"),
         ("46181-0001", (16, 21), "en"),
         ("51000-0010", (1572, 21), "en"),
-        ("61111-0021", (960, 17), "en"),
+        # ("61111-0021", (960, 17), "en"),
         ("63121-0001", (210, 21), "en"),
         ("71311-0001", (600, 25), "en"),
         ("91111-0001", (4719, 17), "en"),
@@ -77,8 +77,8 @@ def test_get_data(
     "table_name, expected_shape",
     [
         ("52111-0001", (68, 22)),
-        ("12211-Z-11", (2152, 14)),
-        ("1000A-2022", (1360, 22)),
+        ("12211-Z-11", (2200, 18)),
+        # ("1000A-2022", (1360, 22)), # metadata request broken
     ],
 )
 def test_get_data_with_quality_on_and_prettify_false(
@@ -90,13 +90,8 @@ def test_get_data_with_quality_on_and_prettify_false(
 
     assert table.data.shape == expected_shape
 
-    if table_name in ["1000A-2022", "52111-0001"]:
-        # check that at least one raw column ends with "_q" for zensus + quality
-        assert any(column.endswith("value_q") for column in table.data.columns)
-    elif table_name == "12211-Z-11":
-        # check that at least no raw column ends with "__q" for regio + quality
-        # (API call with quality="on" does not support quality for regio tables)
-        assert not any(column.endswith("__q") for column in table.data.columns)
+    # check that at least one raw column ends with "_q" for zensus + quality
+    assert any(column.endswith("value_q") for column in table.data.columns)
 
 
 @pytest.mark.vcr()
@@ -105,10 +100,9 @@ def test_get_data_with_quality_on_and_prettify_false(
     [
         (
             "52111-0001",
-            (68, 6),
+            (68, 5),
             (
                 "Jahr",
-                "Deutschland insgesamt",
                 "Beschäftigtengrößenklassen",
                 "WZ2008 (Abschnitte): URS",
                 "Unternehmen (EU)__Anzahl",
@@ -117,28 +111,29 @@ def test_get_data_with_quality_on_and_prettify_false(
         ),
         (
             "12211-Z-11",
-            (2152, 5),
+            (2200, 6),
             (
                 "Jahr",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Kreise und kreisfreie Städte",
                 "Art der Lebensform",
-                "Lebensformen am Hauptwohnort__1000",
+                "Lebensformen__1000",
+                "Lebensformen__1000__q",
             ),
         ),
-        (
-            "1000A-2022",
-            (1360, 7),
-            (
-                "Stichtag",
-                "Amtlicher Regionalschlüssel (ARS)",
-                "Bundesländer",
-                "Alter (10er-Jahresgruppen)",
-                "Einwanderungsgeschichte (ausführlich)",
-                "Personen__Anzahl",
-                "Personen__Anzahl__q",
-            ),
-        ),
+        # (
+        #     "1000A-2022",
+        #     (1360, 7),
+        #     (
+        #         "Stichtag",
+        #         "Amtlicher Regionalschlüssel (ARS)",
+        #         "Bundesländer",
+        #         "Alter (10er-Jahresgruppen)",
+        #         "Einwanderungsgeschichte (ausführlich)",
+        #         "Personen__Anzahl",
+        #         "Personen__Anzahl__q",
+        #     ),
+        # ),
     ],
 )
 def test_get_data_with_quality_on_and_prettify_true(
@@ -161,10 +156,9 @@ def test_get_data_with_quality_on_and_prettify_true(
     [
         (
             "12211-0001",
-            (45, 9),
+            (45, 8),
             (
                 "Jahr",
-                "Deutschland insgesamt",
                 "Geschlecht",
                 "Altersgruppen (u15-75m)",
                 "Bevölkerung in Hauptwohnsitzhaushalten__1000",
@@ -180,8 +174,8 @@ def test_get_data_with_quality_on_and_prettify_true(
             (384, 5),
             (
                 "Stichtag",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Bundesländer",
                 "Geschlecht",
                 "Sozialvers.pflichtig Beschäftigte am Arbeitsort__Anzahl",
             ),
@@ -189,10 +183,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "23111-0001",
-            (33, 10),
+            (33, 9),
             (
                 "Jahr",
-                "Deutschland insgesamt",
                 "Berechnungs-/Belegungstage__1000",
                 "Betten je 100.000 Einwohner__Anzahl",
                 "Betten__Anzahl",
@@ -209,10 +202,10 @@ def test_get_data_with_quality_on_and_prettify_true(
             (4352, 7),
             (
                 "Jahr",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
+                "Amtlicher Gemeindeschlüssel (AGS)",
                 "Quartale",
                 "Herkunfts-Bundesland oder Ausland",
-                "Amtlicher Gemeindeschlüssel (AGS)",
-                "Bundesländer",
                 "Familienstand",
                 "Schwangerschaftsabbrüche__Anzahl",
             ),
@@ -220,10 +213,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "32161-0003",
-            (14, 8),
+            (14, 7),
             (
                 "Jahr",
-                "Deutschland insgesamt",
                 "WZ2008 (2-Steller): Abfallerzeugung",
                 "Beschäftigte__1000",
                 "Betriebe__Anzahl",
@@ -235,11 +227,11 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "32421-0012",
-            (560, 7),
+            (1120, 7),
             (
                 "Jahr",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Bundesländer",
                 "Stoffgruppen",
                 "Einsatzbereiche",
                 "Verwendung klimawirksamer Stoffe (CO2-Äquivalente)__1000 t",
@@ -249,10 +241,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "46181-0001",
-            (8, 6),
+            (8, 5),
             (
                 "Jahr",
-                "Deutschland insgesamt",
                 "Verkehrsart",
                 "Hauptverkehrsverbindung",
                 "Beförderte Personen__Anzahl",
@@ -262,10 +253,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "51000-0010",
-            (262, 10),
+            (262, 9),
             (
                 "Jahr",
-                "Deutschland insgesamt",
                 "Länderverzeichnis für die Außenhandelsstatistik",
                 "Warenverzeichnis Außenhandelsstatistik (4-Steller)",
                 "Ausfuhr: Gewicht__t",
@@ -277,24 +267,24 @@ def test_get_data_with_quality_on_and_prettify_true(
             ),
             "de",
         ),
-        (
-            "61111-0021",
-            (960, 5),
-            (
-                "Jahr",
-                "Monate",
-                "Amtlicher Gemeindeschlüssel (AGS)",
-                "Bundesländer",
-                "Index der Nettokaltmieten__2020=100",
-            ),
-            "de",
-        ),
+        # (
+        #     "61111-0021",
+        #     (960, 5),
+        #     (
+        #         "Jahr",
+        #         "Monate",
+        #         "Amtlicher Gemeindeschlüssel (AGS)__Code",
+        #         "Amtlicher Gemeindeschlüssel (AGS)",
+        #         "Bundesländer",
+        #         "Index der Nettokaltmieten__2020=100",
+        #     ),
+        #     "de",
+        # ),
         (
             "63121-0001",
-            (180, 7),
+            (180, 6),
             (
                 "Jahr",
-                "Deutschland insgesamt",
                 "Haushaltsgröße",
                 "Einnahme- und Ausgabearten",
                 "Durchschnittsbetrag je Haushalt und Monat__EUR",
@@ -305,10 +295,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "71311-0001",
-            (600, 6),
+            (600, 5),
             (
                 "Stichtag zum Quartalsende",
-                "Deutschland insgesamt",
                 "Ebenen des öffentlichen Gesamthaushalts",
                 "Haushalte",
                 "Schuldenarten",
@@ -318,10 +307,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "91111-0001",
-            (4719, 4),
+            (4719, 3),
             (
                 "Jahr",
-                "Deutschland insgesamt",
                 "Indikatoren: Nachhaltige Entwicklung",
                 "Indikatoren__jew. ME",
             ),
@@ -329,11 +317,11 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "11111-02-01-4",
-            (538, 4),
+            (550, 4),
             (
                 "Stichtag",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Kreise und kreisfreie Städte",
                 "Zahl der Gemeinden__Anzahl",
             ),
             "de",
@@ -343,8 +331,8 @@ def test_get_data_with_quality_on_and_prettify_true(
             (44010, 7),
             (
                 "Semester",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Kreise und kreisfreie Städte",
                 "Geschlecht",
                 "Nationalität (inkl. insgesamt)",
                 "Fächergruppe (mit Insgesamt)",
@@ -354,66 +342,66 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "32121-01-02-4",
-            (3766, 5),
+            (3850, 5),
             (
                 "Jahr",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Kreise und kreisfreie Städte",
                 "Abfallarten von Haushaltsabfällen",
-                "Aufkommen an Haushaltsabfaellen (o.E-altgeraete)__t",
+                "Aufkommen an Haushaltsabfällen (oh.Elektroaltger.)__t",
             ),
             "de",
         ),
         (
             "41312-01-01-4",
-            (5918, 5),
+            (6050, 5),
             (
                 "Stichtag",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Kreise und kreisfreie Städte",
                 "Tierarten",
                 "Tiere__Anzahl",
             ),
             "de",
         ),
-        (
-            "52411-02-01-4",
-            (538, 6),
-            (
-                "Jahr",
-                "Amtlicher Gemeindeschlüssel (AGS)",
-                "Kreise und kreisfreie Städte",
-                "Insolvenzverfahren (Unternehmen)__MeasureUnitNotFound!",
-                "Arbeitnehmer__Anzahl",
-                "voraussichtliche Forderungen (Unternehmen)__Tsd._EUR",
-            ),
-            "de",
-        ),
+        # (
+        #     "52411-02-01-4",
+        #     (538, 6),
+        #     (
+        #         "Jahr",
+        #         "Amtlicher Gemeindeschlüssel (AGS)__Code",
+        #         "Amtlicher Gemeindeschlüssel (AGS)",
+        #         "Insolvenzverfahren (Unternehmen)__MeasureUnitNotFound!",
+        #         "Arbeitnehmer__Anzahl",
+        #         "voraussichtliche Forderungen (Unternehmen)__Tsd._EUR",
+        #     ),
+        #     "de",
+        # ),
         (
             "61511-01-03-4",
-            (1076, 8),
+            (1100, 8),
             (
                 "Jahr",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Kreise und kreisfreie Städte",
                 "Baulandverkäufe",
-                "Veraeusserungsfaelle von Bauland__Anzahl",
-                "Veraeusserte Baulandflaeche__1000_qm",
-                "Kaufsumme__Tsd._EUR",
                 "Durchschnittlicher Kaufwert je qm__EUR",
+                "Kaufsumme__Tsd. EUR",
+                "Veräußerte Baulandfläche__1000 qm",
+                "Veräußerungsfälle von Bauland__Anzahl",
             ),
             "de",
         ),
         (
             "73111-01-01-4",
-            (538, 6),
+            (550, 6),
             (
                 "Jahr",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Kreise und kreisfreie Städte",
+                "Gesamtbetrag der Einkünfte__Tsd. EUR",
+                "Lohn- und Einkommensteuer__Tsd. EUR",
                 "Lohn- und Einkommensteuerpflichtige__Anzahl",
-                "Gesamtbetrag der Einkuenfte__Tsd._EUR",
-                "Lohn- und Einkommensteuer__Tsd._EUR",
             ),
             "de",
         ),
@@ -422,24 +410,24 @@ def test_get_data_with_quality_on_and_prettify_true(
             (2736, 7),
             (
                 "Jahr",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Bundesländer",
                 "Haushaltsabfälle",
-                "Haushaltsabfaelle__1000_t",
-                "Haushaltsabfaelle, Index (2010=100)__2010=100",
-                "Haushaltsabfaelle, Anteil an Deutschland__Prozent",
+                "Haushaltsabfälle, Anteil an Deutschland__Prozent",
+                "Haushaltsabfälle, Index (2010=100)__2010=100",
+                "Haushaltsabfälle__1000 t",
             ),
             "de",
         ),
         (
             "AI-N-01-2-5",
-            (13922, 5),
+            (13564, 5),
             (
                 "Jahr",
+                "Amtlicher Gemeindeschlüssel (AGS)__Code",
                 "Amtlicher Gemeindeschlüssel (AGS)",
-                "Gemeinden",
-                "Anteil Siedlungs- und Verkehrsflaeche an Gesamtflaeche__Prozent",
-                "Veraenderung der Siedlungs- und Verkehrsflaeche__Prozent",
+                "Anteil Siedlungs- und Verkehrsfläche an Gesamtfläche__Prozent",
+                "Veränderung der Siedlungs- und Verkehrsfläche__Prozent",
             ),
             "de",
         ),
@@ -448,18 +436,17 @@ def test_get_data_with_quality_on_and_prettify_true(
             (10787, 4),
             (
                 "Stichtag",
+                "Amtlicher Regionalschlüssel (ARS)__Code",
                 "Amtlicher Regionalschlüssel (ARS)",
-                "Gemeinden (Gebietsstand 15.05.2022)",
                 "Personen__Anzahl",
             ),
             "de",
         ),
         (
             "2000S-2003",
-            (72, 5),
+            (72, 4),
             (
                 "Stichtag",
-                "Deutschland",
                 "Einwanderungsgeschichte (ausführlich)",
                 "Erwerbsstatus",
                 "Personen__Anzahl",
@@ -468,10 +455,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "3000G-1008",
-            (10, 5),
+            (10, 4),
             (
                 "Stichtag",
-                "Deutschland",
                 "Energieträger der Heizung",
                 "Gebäude mit Wohnraum__%",
                 "Gebäude mit Wohnraum__Anzahl",
@@ -480,10 +466,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "4000W-2030",
-            (286, 5),
+            (286, 4),
             (
                 "Stichtag",
-                "Deutschland",
                 "Miete der Wohnung (100 €-Schritte)",
                 "Gebäudetyp (Größe)",
                 "Vermietete Wohnungen in Gebäuden mit Wohnraum__Anzahl",
@@ -492,10 +477,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "81000-0001",
-            (40, 10),
+            (40, 9),
             (
                 "Jahr",
-                "Deutschland insgesamt",
                 "Preisbasis (jeweilige Preise / preisbereinigt)",
                 "Bruttoinlandsprodukt (Veränderung in %)__Prozent",
                 "Bruttoinlandsprodukt je Einwohner__jew. ME",
@@ -509,10 +493,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "12211-0001",
-            (45, 9),
+            (45, 8),
             (
                 "Year",
-                "Germany",
                 "Sex",
                 "Age groups (under 15 - 75 years and over)",
                 "Economically active population from prim.resid.hh.__1000",
@@ -528,8 +511,8 @@ def test_get_data_with_quality_on_and_prettify_true(
             (384, 5),
             (
                 "Reference date",
+                "Official municipality key (AGS)__Code",
                 "Official municipality key (AGS)",
-                "Länder",
                 "Sex",
                 "Employees subj. to social insur. at place of work__number",
             ),
@@ -537,10 +520,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "23111-0001",
-            (33, 10),
+            (33, 9),
             (
                 "Year",
-                "Germany",
                 "Average length of stay__days",
                 "Average occupancy of hospital beds__percent",
                 "Beds per 100 000 inhabitants__number",
@@ -557,10 +539,10 @@ def test_get_data_with_quality_on_and_prettify_true(
             (4352, 7),
             (
                 "Year",
+                "Official municipality key (AGS)__Code",
+                "Official municipality key (AGS)",
                 "Quarters",
                 "Land of origin or origin from abroad",
-                "Official municipality key (AGS)",
-                "Länder",
                 "Marital status",
                 "Terminations of pregnancy__number",
             ),
@@ -568,10 +550,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "32161-0003",
-            (14, 8),
+            (14, 7),
             (
                 "Year",
-                "Germany",
                 "WZ2008 (2-digit codes): Waste production",
                 "Coverage of local units__percent",
                 "Coverage of persons employed__percent",
@@ -583,11 +564,11 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "32421-0012",
-            (560, 7),
+            (1120, 7),
             (
                 "Year",
+                "Official municipality key (AGS)__Code",
                 "Official municipality key (AGS)",
-                "Länder",
                 "Groups of substances",
                 "Application areas",
                 "Use of climate-affecting substances (CO2-equiv.)__1000 t",
@@ -597,10 +578,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "46181-0001",
-            (8, 6),
+            (8, 5),
             (
                 "Year",
-                "Germany",
                 "Type of transport",
                 "Main traffic relation",
                 "Passengers carried__number",
@@ -610,10 +590,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "51000-0010",
-            (262, 10),
+            (262, 9),
             (
                 "Year",
-                "Germany",
                 "Country Nomenclature for External Trade Statistics",
                 "Commodity Class. Foreign Trade Stat.(4-digit cod.)",
                 "Exports: Net mass__t",
@@ -625,24 +604,23 @@ def test_get_data_with_quality_on_and_prettify_true(
             ),
             "en",
         ),
-        (
-            "61111-0021",
-            (960, 5),
-            (
-                "Year",
-                "Months",
-                "Official municipality key (AGS)",
-                "Länder",
-                "Index of net rents exclusive of heating expenses__2020=100",
-            ),
-            "en",
-        ),
+        # (
+        #     "61111-0021",
+        #     (960, 5),
+        #     (
+        #         "Year",
+        #         "Months",
+        #         "Official municipality key (AGS)__Code",
+        #         "Official municipality key (AGS)",
+        #         "Index of net rents exclusive of heating expenses__2020=100",
+        #     ),
+        #     "en",
+        # ),
         (
             "63121-0001",
-            (180, 7),
+            (180, 6),
             (
                 "Year",
-                "Germany",
                 "Household size",
                 "Types of income and expenditure",
                 "Average amount per household and month__EUR",
@@ -653,10 +631,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "71311-0001",
-            (600, 6),
+            (600, 5),
             (
                 "Reference date end-of-quarter",
-                "Germany",
                 "Levels of the overall public budget",
                 "Budgets",
                 "Types of debts",
@@ -666,10 +643,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "91111-0001",
-            (4719, 4),
+            (4719, 3),
             (
                 "Year",
-                "Germany",
                 "Indicators: Sustainable development",
                 "Indicators__unit app.",
             ),
@@ -680,18 +656,17 @@ def test_get_data_with_quality_on_and_prettify_true(
             (10787, 4),
             (
                 "Reference date",
+                "Official regional key (ARS)__Code",
                 "Official regional key (ARS)",
-                "Municipalities (territory on 15 May 2022, LAU-2)",
                 "Persons__number",
             ),
             "en",
         ),
         (
             "2000S-2003",
-            (72, 5),
+            (72, 4),
             (
                 "Reference date",
-                "Germany",
                 "Immigration history (in detail)",
                 "Activity status",
                 "Persons__number",
@@ -700,10 +675,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "3000G-1008",
-            (10, 5),
+            (10, 4),
             (
                 "Reference date",
-                "Germany",
                 "Energy source used for heating",
                 "Buildings with residential space__%",
                 "Buildings with residential space__number",
@@ -712,10 +686,9 @@ def test_get_data_with_quality_on_and_prettify_true(
         ),
         (
             "4000W-2041",
-            (90, 6),
+            (90, 5),
             (
                 "Reference date",
-                "Germany",
                 "Duration of dwelling vacancy",
                 "Floor area of the dwelling (10m² increments)",
                 "Vacant dwellings in buildings with residential space__%",
@@ -746,7 +719,18 @@ def test_prettify(
 @pytest.mark.vcr()
 @pytest.mark.parametrize(
     "table_name, time_col, language",
-    [("12411-01-01-4", "Stichtag", "de"), ("12411-01-01-4", "Stichtag", "en")],
+    [
+        ("12411-01-01-4", "Stichtag", "de"),
+        ("12411-01-01-4", "Stichtag", "en"),
+        ("13111-0005", "Stichtag", "de"),
+        ("13111-0005", "Reference date", "en"),
+        ("71311-0001", "Stichtag zum Quartalsende", "de"),
+        ("71311-0001", "Reference date end-of-quarter", "en"),
+        ("1000A-0000", "Stichtag", "de"),
+        ("1000A-0000", "Reference date", "en"),
+        ("2000S-2003", "Stichtag", "de"),
+        ("2000S-2003", "Reference date", "en"),
+    ],
 )
 def test_dtype_time_column(mocker, table_name: str, time_col: str, language: str):
     mocker.patch.object(pystatis.db, "check_credentials", return_value=True)